project/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-cpf>=0.1.0",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
    "rest-framework-simplejwt>=0.0.2",
]
